requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
fuzzywuzzy==0.18.0
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
